requirements.txt
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
asgiref
attrs
certifi
chardet
charset-normalizer
cloudinary
crispy-bootstrap5
distlib
dj-database-url
dj-rest-auth
Django
django-allauth
django-cloudinary-storage
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
djangorestframework-simplejwt
filelock
frozenlist
gunicorn
idna
Markdown
MouseInfo
multidict
mysql-connector-python
packaging
pillow
platformdirs
propcache
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyJWT
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
pytz
reportlab
requests
setuptools
six
sqlparse
twilio
typing_extensions
tzdata
urllib3
virtualenv
wheel
whitenoise
yarl
